CATO-3000  Moved common retriever functions for repayments to their own trait.
This is so we can use them independently of the rest of CT600, which is required for a charity return with all proceeds used for charitable purposes.
